HUD Improvements, 1.16 port, fix NanoVG with ARM (#52)
* egg 1

* separate Hud from background stuff

* 1984

This reverts commit 9ae517d57bbd495d30d35cb1cbfe81a03556e6bd.

* hitboxes woo!!!!!

* Revert "hitboxes woo!!!!!"

This reverts commit 405d32d17df3c83f2e79eddf0de853f7279767a6.

* padding

* allow position to go slightly off the screen

* stop using ints for ABSOLUTELY EVERYTHING, DIAMOND ...
fix vigilance compat not setting color

* start on new pos system

* some stuff

* finish new position system

* api momento

* 1.16.2 fabric port

* start on hud gui

* temp remove 1.16.2 fabric since it doesn't compile

* fix fabric build

* hud gui stuff

* apiDump

* fix fabric build 2

* so true

* selecting stuff

* scaling + other small things

* More protecting

* fix nanovg not working with macOS ARM
move OneConfig.preLaunch to OneConfigInit

* clean up OneUIScreen
make kotlin version of TestNanoVGGui

* make keybinds have runnable by default

* rollback keybind things

* merge master into hud-improvements (#55)

* Release workflow (#53)

* release workflow

* update normal version to hash

* fix

* fix naming

* fix some stuff

* fix version thing

* switch to number from hash

* Release workflow (#54)

* release workflow

* update normal version to hash

* fix

* fix naming

* fix some stuff

* fix version thing

* switch to number from hash

* Maybe epic fixo

* gotta love those Java principles

* Revert "gotta love those Java principles", wrong branch

This reverts commit 333d8b2ad8941790c13c4bfe0777fbd203d463e5.

* start on snapping

* Finish snapping

* stop including mixin by default on legacy versions
this breaks builds if the mod itself does not use mixin

* merge draw and drawExample

* fix gradle publish

* Some fixes

* Api DUmpidy

* Help subcommand impovments (#59)

* Made the overall look of the "help" subcommand better + added the ability to change the colour for the command overall + each individual SubCommand

* Made the alliases show batter + added support for to show subcommand aliasses

* mr deliverer didnt reply but whatever, added a space between command/subcommand and alliasses


* fix file not overwriting
toJavaColor

* Fix full shadow not scaling correctly

+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/BasicButton.java
@@ -16,9 +16,9 @@ public class BasicButton extends BasicElement {
     protected SVGs icon1, icon2;
     private final int alignment;
     private final float fontSize, cornerRadius;
-    private final int xSpacing, xPadding;
+    private final float xSpacing, xPadding;
     private final int iconSize;
-    public int x, y;
+    public float x, y;
     public static final int ALIGNMENT_LEFT = 0;
     public static final int ALIGNMENT_CENTER = 2;
     public static final int ALIGNMENT_JUSTIFIED = 3;
@@ -57,7 +57,7 @@ public class BasicButton extends BasicElement {
     }
 
     @Override
-    public void draw(long vg, int x, int y) {
+    public void draw(long vg, float x, float y) {
         this.x = x;
         this.y = y;
         this.update(x, y);
