+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/ColorSelector.java
@@ -45,21 +45,21 @@ public class ColorSelector {
     private final ColorSlider topSlider = new ColorSlider(384, 0, 360, 127);
     private final ColorSlider bottomSlider = new ColorSlider(384, 0, 255, 100);
     private final Slider speedSlider = new Slider(296, 1, 32, 0);
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private Animation barMoveAnimation = new DummyAnimation(18);
     private Animation moveAnimation = new DummyAnimation(1);
-    private int mouseX, mouseY;
+    private float mouseX, mouseY;
     private int mode = 0;
     private boolean dragging, mouseWasDown;
     private final boolean hasAlpha;
     private Scissor inputScissor = null;
 
-    public ColorSelector(OneColor color, int mouseX, int mouseY) {
+    public ColorSelector(OneColor color, float mouseX, float mouseY) {
         this(color, mouseX, mouseY, true);
     }
 
-    public ColorSelector(OneColor color, int mouseX, int mouseY, boolean hasAlpha) {
+    public ColorSelector(OneColor color, float mouseX, float mouseY, boolean hasAlpha) {
         this.color = color;
         this.hasAlpha = hasAlpha;
         buttons.add(new BasicButton(124, 28, "HSB Box", BasicButton.ALIGNMENT_CENTER, ColorPalette.TERTIARY));
@@ -82,12 +82,12 @@ public class ColorSelector {
         this.y = Math.max(0, mouseY - 776);
         if (color.getDataBit() != -1) mode = 2;
         if (mode == 0 || mode == 2) {
-            this.mouseX = (int) (color.getSaturation() / 100f * 384 + x + 16);
-            this.mouseY = (int) (Math.abs(color.getBrightness() / 100f - 1f) * 288 + y + 120);
+            this.mouseX = (color.getSaturation() / 100f * 384 + x + 16);
+            this.mouseY = (Math.abs(color.getBrightness() / 100f - 1f) * 288 + y + 120);
         } else {
             topSlider.setValue(color.getBrightness() / 100f * 360f);
-            this.mouseX = (int) (Math.sin(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + x + 208);
-            this.mouseY = (int) (Math.cos(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + y + 264);
+            this.mouseX = (float) (Math.sin(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + x + 208);
+            this.mouseY = (float) (Math.cos(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + y + 264);
         }
         //for(OneColor color1 : OneConfigConfig.recentColors) {
         //    recentColors.add(new ColorBox(color1));
@@ -178,7 +178,7 @@ public class ColorSelector {
 
         setColorFromXY();
         if (mode != 2) color.setChromaSpeed(-1);
-        drawColorSelector(vg, mode, (int) (x * percentMoveMain), y);
+        drawColorSelector(vg, mode, (x * percentMoveMain), y);
         if (dragging && InputUtils.isClicked(true)) {
             dragging = false;
         }
@@ -209,7 +209,7 @@ public class ColorSelector {
         }
     }
 
-    private void drawColorSelector(long vg, int mode, int x, int y) {
+    private void drawColorSelector(long vg, int mode, float x, float y) {
         switch (mode) {
             default:
             case 0:
@@ -242,8 +242,8 @@ public class ColorSelector {
 
     private void doDrag() {
         if (InputUtils.isAreaHovered(x, y, 368, 64) && Platform.getMousePlatform().isButtonDown(0) && !dragging) {
-            int dx = (int) (Platform.getMousePlatform().getMouseDX() / (OneConfigGui.INSTANCE == null ? 1 : OneConfigGui.INSTANCE.getScaleFactor()));
-            int dy = (int) (Platform.getMousePlatform().getMouseDY() / (OneConfigGui.INSTANCE == null ? 1 : OneConfigGui.INSTANCE.getScaleFactor()));
+            float dx = (float) (Platform.getMousePlatform().getMouseDX() / (OneConfigGui.INSTANCE == null ? 1 : OneConfigGui.INSTANCE.getScaleFactor()));
+            float dy = (float) (Platform.getMousePlatform().getMouseDY() / (OneConfigGui.INSTANCE == null ? 1 : OneConfigGui.INSTANCE.getScaleFactor()));
             x += dx;
             mouseX += dx;
             y -= dy;
@@ -279,8 +279,8 @@ public class ColorSelector {
                 }
                 break;
             case 1:
-                int circleCenterX = x + 208;
-                int circleCenterY = y + 264;
+                float circleCenterX = x + 208;
+                float circleCenterY = y + 264;
                 double squareDist = Math.pow((circleCenterX - InputUtils.mouseX()), 2) + Math.pow((circleCenterY - InputUtils.mouseY()), 2);
                 hovered = squareDist < 144 * 144 && Platform.getMousePlatform().isButtonDown(0);
                 isMouseDown = Platform.getMousePlatform().isButtonDown(0);
@@ -293,8 +293,8 @@ public class ColorSelector {
                     if (angle < 0) angle += 360;
                     if ((squareDist / (144 * 144) > 1f)) {
                         saturation = 100;
-                        mouseX = (int) (Math.sin(Math.toRadians(-angle) + 1.5708) * 144 + x + 208);
-                        mouseY = (int) (Math.cos(Math.toRadians(-angle) + 1.5708) * 144 + y + 264);
+                        mouseX = (float) (Math.sin(Math.toRadians(-angle) + 1.5708) * 144 + x + 208);
+                        mouseY = (float) (Math.cos(Math.toRadians(-angle) + 1.5708) * 144 + y + 264);
                     } else {
                         saturation = (int) (squareDist / (144 * 144) * 100);
                         mouseX = InputUtils.mouseX();
@@ -309,14 +309,14 @@ public class ColorSelector {
     private void setXYFromColor() {
         bottomSlider.setValue(color.getAlpha());
         if (mode == 1) {
-            mouseX = (int) (Math.sin(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + x + 208);
-            mouseY = (int) (Math.cos(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + y + 264);
+            mouseX = (float) (Math.sin(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + x + 208);
+            mouseY = (float) (Math.cos(Math.toRadians(-color.getHue()) + 1.5708) * (saturationInput.getCurrentValue() / 100 * 144) + y + 264);
             topSlider.setValue(color.getBrightness() / 100f * 360f);
         }
         if (mode == 0 || mode == 2) {
             topSlider.setValue(color.getHue());
-            mouseX = (int) (saturationInput.getCurrentValue() / 100f * 384 + x + 16);
-            mouseY = (int) (Math.abs(brightnessInput.getCurrentValue() / 100f - 1f) * 288 + y + 120);
+            mouseX = (saturationInput.getCurrentValue() / 100f * 384 + x + 16);
+            mouseY = (Math.abs(brightnessInput.getCurrentValue() / 100f - 1f) * 288 + y + 120);
         }
     }
 
@@ -429,7 +429,7 @@ public class ColorSelector {
         }
 
         @Override
-        public void draw(long vg, int x, int y) {
+        public void draw(long vg, float x, float y) {
             if (!disabled) update(x, y);
             else RenderManager.setAlpha(vg, 0.5f);
             super.dragPointerSize = 15f;
@@ -469,7 +469,7 @@ public class ColorSelector {
         }
 
         @Override
-        public void draw(long vg, int x, int y) {
+        public void draw(long vg, float x, float y) {
             RenderManager.drawRoundedRect(vg, x, y, 32, 32, toggled ? Colors.PRIMARY_600 : Colors.GRAY_300, 12f);
             RenderManager.drawRoundedRect(vg, x + 2, y + 2, 28, 28, Colors.GRAY_800, 10f);
             RenderManager.drawRoundedRect(vg, x + 4, y + 4, 24, 24, color.getRGB(), 8f);
